requirements.txt
easyocr
opencv-python
pymupdf
pillow
flask